# LocalAIAgent.API/openapi/LocalAIAgent.API.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LocalAIAgent.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/admin/invitations/{invitationId}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/users/{userId}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManagedUserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManagedUserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManagedUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Feeds/Toggle": {
      "post": {
        "tags": [
          "Feeds"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleFeedDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleFeedDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleFeedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Feeds/Custom": {
      "post": {
        "tags": [
          "Feeds"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCustomFeedDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCustomFeedDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddCustomFeedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FeedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Feeds/Custom/{customFeedId}": {
      "delete": {
        "tags": [
          "Feeds"
        ],
        "parameters": [
          {
            "name": "customFeedId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddCustomFeedDto": {
        "required": [
          "displayName",
          "language",
          "urls"
        ],
        "type": "object",
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FeedDto": {
        "required": [
          "clientName",
          "displayName",
          "enabled",
          "isCustom",
          "language",
          "languageName"
        ],
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "languageName": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "isCustom": {
            "type": "boolean"
          },
          "customFeedId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "lastFetchErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ToggleFeedDto": {
        "required": [
          "clientName",
          "enabled"
        ],
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateManagedUserDto": {
        "type": "object",
        "properties": {
          "isDisabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}